nvk: Advertise minStorageBufferOffsetAlignment=4 for VKD3D
VKD3D won't enable the VK_EXT_descriptor_buffer path unless we advertise
a minStorageBufferOffsetAlignment of 4.  On the NVIDIA proprietary
driver, they get around the limit by just detecting the driver and
smashing it to 4 internally.  It's better if we set the limit for them
so that we can also adjust NIR's optimizations to not assume higher
alignments.

v2 (mhenning): Rebase and always initialize value of
               instance->ssbo_align_4b

Reviewed-by: Mel Henning <mhenning@darkrefraction.com>